pyproject.toml
[project]
name = "plenary-pantry"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "chromedriver>=2.24.1",
    "django>=5.2.5",
    "django-extensions>=4.1",
    "nltk>=3.9.1",
    "opencv-python-headless>=4.12.0.88",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.0.0",
    "pytest-django>=4.8.0",
    "pytesseract>=0.3.13",
    "python-decouple>=3.8",
    "requests>=2.32.4",
    "selenium>=4.34.2",
    "spacy>=3.8.7",
    "tesseract>=0.1.3",
    "pytest-cov>=6.2.1",
    "coverage>=7.10.6",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "plenary_pantry.test_settings"
django_find_project = false
python_files = ["tests.py", "test_*.py", "*_tests.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
pythonpath = ["."]
addopts = ["--tb=short", "--strict-markers", "--import-mode=importlib"]
testpaths = ["recipe_ingestion/tests"]
norecursedirs = [".git", ".tox", "dist", "build", "*.egg", "recipe_ingestion/management", "recipe_ingestion/migrations", "recipe_ingestion/__pycache__"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
